+++ script/bootstrap
@@ -12,9 +12,6 @@ if [ -f "Brewfile" ] && [ "$(uname -s)" = "Darwin" ]; then
     echo "==> Installing Homebrew dependencies…"
     brew bundle
   }
-
-  echo "==> Setup nginx config"
-  brew setup-nginx-conf osumbchallenges . config/nginx.conf.erb
 fi
 
 npm install
